feat: changed logs by color feedback
The previous logs section has been removed because it was useless.
We'd just two messages to show: success and failure. Now we're using
a section dedicated to manage the background of the mainActivity. when
data sending was donde correctly the background turns green and when
something went wrong, then it turns red.

refs: #miot-66